Add files via upload
The core of the JS code in this project was created in response to the Roman Numeral algorithm challenge at FCC. I created the html and css files after that code was done and then had to adapt the code to work with the site.  Upon first upload the site was slightly buggy and thanks are owed to Lionel Rowe who helped me see the solution while increasing my understanding of JavaScript.